backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
akshare==1.16.20
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.1
ta==0.11.0
scipy==1.15.0
httpx==0.28.1
google-genai==1.5.0
anthropic==0.43.0
openai==1.58.1
tavily-python==0.5.0
aiosqlite==0.20.0
pytest==8.3.4
pytest-asyncio==0.25.0
